+++ resources/views/index.blade.php
@@ -109,7 +109,7 @@
         </div>
         <div class="col-sm-6">
             <h2 class="featurette-heading">Configure to your needs</h2>
-            <p class="lead">StyleCI is powered by <a href="https://github.com/FriendsOfPHP/PHP-CS-Fixer" target="_blank">PHP CS Fixer</a>, and coming fully configurable to fix PSR-1, PSR-2, PSR-5, Symfony and various other checks via a <code>.styleci.yml</code> file, committed to the root your repository. Coming soon!</p>
+            <p class="lead">StyleCI is powered by <a href="https://github.com/FriendsOfPHP/PHP-CS-Fixer" target="_blank">PHP CS Fixer</a>, and coming fully configurable to fix PSR-1, PSR-2, PSR-5, Symfony and various other checks via a <code>.styleci.yml</code> file, committed to the root your repository.</p>
         </div>
     </div>
 </div>